backend/pyproject.toml
[project]
name = "flight-booking-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.1",
    "fastapi[all]>=0.115.11",
    "passlib[bcrypt]>=1.7.4",
    "psycopg[binary]>=3.2.5",
    "pyjwt>=2.10.1",
    "sqlmodel>=0.0.24",
]

[dependency-groups]
dev = [
    "coverage>=7.6.12",
    "mypy>=1.15.0",
    "pytest>=8.3.5",
    "ruff>=0.9.10",
    "types-passlib>=1.7.7.20241221",
]

[tool.mypy]
strict = true
exclude = ["migrations"]

[tool.ruff]
target-version = "py312"
exclude = ["migrations"]

[tool.ruff.lint]
select = [
    "F",  # Pyflakes
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "I",  # isort
    "N",  # pep8-naming
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
